Merp fuxes
Fuxed the emergency armory door not working for all access by replacing it with not a captains door and renaming it.  I would never do that shit.

Fixed cargo and one of the dorms being off the grid, because they were just not tied to the red line.

Added a few more solar panels, because why the fuck not.  #yolo